# Conflicts:
#	lab1/doublebasepalindromes/CMakeLists.txt
#	lab1/doublebasepalindromes/DoubleBasePalindromes.cpp
#	lab1/doublebasepalindromes/DoubleBasePalindromes.h
#	lab1/doublebasepalindromes/main.cpp
#	lab1/factorial/Factorial.cpp
#	lab1/multiplicationtable/CMakeLists.txt
#	lab1/multiplicationtable/MultiplicationTable.cpp
#	lab1/multiplicationtable/MultiplicationTable.h
#	lab1/multiplicationtable/main.cpp
#	lab1/palindrome/CMakeLists.txt
#	lab1/palindrome/Palindrome.cpp
#	lab1/palindrome/Palindrome.h
#	lab1/palindrome/main.cpp
#	lab1/reversestring/CMakeLists.txt
#	lab1/reversestring/ReverseString.cpp
#	lab1/reversestring/ReverseString.h
#	lab1/reversestring/main.cpp
#	lab2/polybius/Polybius.cpp
#	lab2/polybius/Polybius.h
#	lab2/polybius/main.cpp
#	lab2/xorcypherbreaker/CMakeLists.txt
#	lab2/xorcypherbreaker/XorCypherBreaker.cpp
#	lab2/xorcypherbreaker/XorCypherBreaker.h

+++ lab1/factorial/Factorial.h
@@ -1,10 +1,10 @@
 //
-// Created by mwypych on 02.02.17.
+// Created by jan on 02.02.17.
 //
 
 #ifndef JIMP_EXERCISES_FACTORIAL_H
 #define JIMP_EXERCISES_FACTORIAL_H
 
-int factorial(int value);
+long long int factorial(int value);
 
 #endif //JIMP_EXERCISES_FACTORIAL_H
